requirements.txt
Django==3.2.9
Pillow==9.1.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pyTelegramBotAPI==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
